+++ packages/moonstone/Picker/Picker.js
@@ -5,6 +5,7 @@
  * @module moonstone/Picker
  */
 
+import Changeable from '@enact/ui/Changeable';
 import clamp from 'ramda/src/clamp';
 import kind from '@enact/core/kind';
 import React from 'react';
@@ -190,15 +191,23 @@ const PickerBase = kind({
 /**
  * A Picker component that allows selecting values from a list of values.
  *
+ * By default, `Picker` maintains the state of its `value` property. Supply the `defaultValue`
+ * property to control its initial value. If you wish to directly control updates to the component,
+ * supply a value to `value` at creation time and update it in response to `onChange` events.
+ *
  * @class Picker
  * @memberof moonstone/Picker
+ * @mixes ui/Changeable.Changeable
+ * @mixes moonstone/Marquee.MarqueeController
  * @ui
  * @public
  */
-const Picker = MarqueeController(
-	{marqueeOnFocus: true},
-	SpottablePicker(
-		PickerBase
+const Picker = Changeable(
+	MarqueeController(
+		{marqueeOnFocus: true},
+		SpottablePicker(
+			PickerBase
+		)
 	)
 );