+++ packages/moonstone/RadioItem/RadioItem.js
@@ -6,17 +6,18 @@
 
 import kind from '@enact/core/kind';
 import React, {PropTypes} from 'react';
+import Toggleable from '@enact/ui/Toggleable';
 
-import ToggleItem from '../ToggleItem';
+import {ToggleItemBase} from '../ToggleItem';
 
 import css from './RadioItem.less';
 
 /**
- * {@link moonstone/RadioItem.RadioItem} is a component that
+ * {@link moonstone/RadioItem.RadioItemBase} is a component that
  * combines a Toggleable radio selector and an Item. It has two selected states
  * `true` & `false`.
  *
- * @class RadioItem
+ * @class RadioItemBase
  * @memberof moonstone/RadioItem
  * @ui
  * @public
@@ -24,7 +25,7 @@ import css from './RadioItem.less';
 const RadioItemBase = kind({
 	name: 'RadioItem',
 
-	propTypes: /** @lends moonstone/RadioItem.RadioItem.prototype */ {
+	propTypes: /** @lends moonstone/RadioItem.RadioItemBase.prototype */ {
 		/**
 		 * The string to be displayed as the main content of the radio item.
 		 *
@@ -66,7 +67,6 @@ const RadioItemBase = kind({
 		 * Applies a filled circle icon to the radio item.
 		 *
 		 * @type {Boolean}
-		 * @default false
 		 * @public
 		 */
 		selected: PropTypes.bool,
@@ -84,7 +84,6 @@ const RadioItemBase = kind({
 	defaultProps: {
 		disabled: false,
 		inline: false,
-		selected: false,
 		value: ''
 	},
 
@@ -104,9 +103,31 @@ const RadioItemBase = kind({
 	},
 
 	render: (props) => (
-		<ToggleItem {...props} />
+		<ToggleItemBase {...props} />
 	)
 });
 
-export default RadioItemBase;
-export {RadioItemBase as RadioItem, RadioItemBase};
+
+/**
+ * {@link moonstone/RadioItem.RadioItem} is a component that combines a
+ * {@link ui/Toggleable.Toggleable} radio selector and an Item. It has two selected states `true` &
+ * `false`.
+ *
+ * By default, `RadioItem` maintains the state of its `selected` property. Supply the
+ * `defaultSelected` property to control its initial value. If you wish to directly control updates
+ * to the component, supply a value to `selected` at creation time and update it in response to
+ * `onToggle` events.
+ *
+ * @class RadioItem
+ * @memberof moonstone/RadioItem
+ * @mixes ui/Toggleable.Toggleable
+ * @ui
+ * @public
+ */
+const RadioItem = Toggleable(
+	{prop: 'selected'},
+	RadioItemBase
+);
+
+export default RadioItem;
+export {RadioItem, RadioItemBase};