+++ packages/moonstone/Scroller/Scrollable.js
@@ -6,16 +6,22 @@
 
 import clamp from 'ramda/src/clamp';
 import classNames from 'classnames';
+import {contextTypes} from '@enact/ui/Resizable';
+import {forward} from '@enact/core/handle';
 import {getDirection} from '@enact/spotlight';
 import hoc from '@enact/core/hoc';
 import {Job} from '@enact/core/util';
 import React, {Component, PropTypes} from 'react';
-import {contextTypes} from '@enact/ui/Resizable';
 import ri from '@enact/ui/resolution';
 
 import css from './Scrollable.less';
 import ScrollAnimator from './ScrollAnimator';
 import Scrollbar from './Scrollbar';
+
+const
+	forwardScroll = forward('onScroll'),
+	forwardScrollStart = forward('onScrollStart'),
+	forwardScrollStop = forward('onScrollStop');
 
 const
 	calcVelocity = (d, dt) => (d && dt) ? d / dt : 0,
@@ -57,6 +63,8 @@ const dataIndexAttribute = 'data-index';
  */
 const ScrollableHoC = hoc((config, Wrapped) => {
 	return class Scrollable extends Component {
+		static displayName = 'Scrollable'
+
 		static propTypes = /** @lends moonstone/Scroller.Scrollable.prototype */ {
 			/**
 			 * The callback function which is called for linking scrollTo function.
@@ -79,8 +87,6 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			 * @public
 			 */
 			cbScrollTo: PropTypes.func,
-
-			className: PropTypes.string,
 
 			/**
 			 * Specifies how to show horizontal scrollbar. Acceptable values are `'auto'`,
@@ -140,6 +146,48 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 
 		static childContextTypes = contextTypes
 
+		constructor (props) {
+			super(props);
+
+			this.state = {
+				isHorizontalScrollbarVisible: this.isHorizontalScrollbarVisible(),
+				isVerticalScrollbarVisible: this.isVerticalScrollbarVisible()
+			};
+
+			this.initChildRef = this.initRef('childRef');
+			this.initContainerRef = this.initRef('containerRef');
+
+			const {onKeyDown, onKeyUp} = this;
+			// We have removed all mouse event handlers for now.
+			// Revisit later for touch usage.
+			this.eventHandlers = {
+				onKeyDown,
+				onKeyUp
+			};
+
+			this.verticalScrollbarProps = {
+				ref: this.initRef('scrollbarVerticalRef'),
+				vertical: true,
+				onPrevScroll: this.initScrollbarBtnHandler('vertical', -1),
+				onNextScroll: this.initScrollbarBtnHandler('vertical', 1)
+			};
+
+			this.horizontalScrollbarProps = {
+				ref: this.initRef('scrollbarHorizontalRef'),
+				vertical: false,
+				onPrevScroll: this.initScrollbarBtnHandler('horizontal', -1),
+				onNextScroll: this.initScrollbarBtnHandler('horizontal', 1)
+			};
+
+			props.cbScrollTo(this.scrollTo);
+		}
+
+		getChildContext () {
+			return {
+				invalidateBounds: this.enqueueForceUpdate
+			};
+		}
+
 		// status
 		horizontalScrollability = false
 		verticalScrollability = false
@@ -189,48 +237,6 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		// scroll animator
 		animator = new ScrollAnimator()
 
-		constructor (props) {
-			super(props);
-
-			this.state = {
-				isHorizontalScrollbarVisible: this.isHorizontalScrollbarVisible(),
-				isVerticalScrollbarVisible: this.isVerticalScrollbarVisible()
-			};
-
-			this.initChildRef = this.initRef('childRef');
-			this.initContainerRef = this.initRef('containerRef');
-
-			const {onKeyDown, onKeyUp} = this;
-			// We have removed all mouse event handlers for now.
-			// Revisit later for touch usage.
-			this.eventHandlers = {
-				onKeyDown,
-				onKeyUp
-			};
-
-			this.verticalScrollbarProps = {
-				ref: this.initRef('scrollbarVerticalRef'),
-				vertical: true,
-				onPrevScroll: this.initScrollbarBtnHandler('vertical', -1),
-				onNextScroll: this.initScrollbarBtnHandler('vertical', 1)
-			};
-
-			this.horizontalScrollbarProps = {
-				ref: this.initRef('scrollbarHorizontalRef'),
-				vertical: false,
-				onPrevScroll: this.initScrollbarBtnHandler('horizontal', -1),
-				onNextScroll: this.initScrollbarBtnHandler('horizontal', 1)
-			};
-
-			props.cbScrollTo(this.scrollTo);
-		}
-
-		getChildContext () {
-			return {
-				invalidateBounds: this.enqueueForceUpdate
-			};
-		}
-
 		// handle an input event
 
 		dragStart (e) {
@@ -245,8 +251,9 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		}
 
 		drag (e) {
-			let t = perf.now();
-			const d = this.dragInfo;
+			const
+				t = perf.now(),
+				d = this.dragInfo;
 
 			if (this.horizontalScrollability) {
 				d.dx = e.clientX - d.clientX;
@@ -469,15 +476,15 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		// call scroll callbacks
 
 		doScrollStart () {
-			this.props.onScrollStart({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()});
+			forwardScrollStart({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()}, this.props);
 		}
 
 		doScrolling () {
-			this.props.onScroll({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()});
+			forwardScroll({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()}, this.props);
 		}
 
 		doScrollStop () {
-			this.props.onScrollStop({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()});
+			forwardScrollStop({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()}, this.props);
 		}
 
 		// update scroll position
@@ -644,7 +651,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 
 		scrollTo = (opt) => {
 			if (!this.isInitializing) {
-				let {left, top} = this.getPositionForScrollTo(opt);
+				const {left, top} = this.getPositionForScrollTo(opt);
 				this.scrollToInfo = null;
 
 				if (left !== null || top !== null) {
@@ -804,7 +811,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			this.forceUpdateJob.start();
 		}
 
-		forceUpdateJob = new Job(this.forceUpdate, 32)
+		forceUpdateJob = new Job(this.forceUpdate.bind(this), 32)
 
 		// render
 
@@ -853,11 +860,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					className
 				);
 
-			delete props.className;
 			delete props.cbScrollTo;
+			delete props.className;
+			delete props.horizontalScrollbar;
 			delete props.style;
 			delete props.verticalScrollbar;
-			delete props.horizontalScrollbar;
 
 			return (
 				(isHorizontalScrollbarVisible || isVerticalScrollbarVisible) ? (
@@ -866,7 +873,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 						{hscrollbar}
 						<Wrapped {...props} {...this.eventHandlers} ref={this.initChildRef} cbScrollTo={this.scrollTo} className={css.container} />
 					</div>
-				) : <Wrapped {...props} {...this.eventHandlers} ref={this.initChildRef} cbScrollTo={this.scrollTo} className={scrollableClasses} style={style} />
+				) : <Wrapped {...props} {...this.eventHandlers} cbScrollTo={this.scrollTo} className={scrollableClasses} ref={this.initChildRef} style={style} />
 			);
 		}
 	};