+++ packages/moonstone/Slider/Slider.js
@@ -4,6 +4,7 @@
  * @module moonstone/Slider
  */
 
+import Changeable from '@enact/ui/Changeable';
 import factory from '@enact/core/factory';
 import {forKey, forward, handle, stopImmediate} from '@enact/core/handle';
 import kind from '@enact/core/kind';
@@ -124,16 +125,6 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			min: PropTypes.number,
 
 			/**
-			 * Disables the built-in tooltip. A custom tooltip may still be used by supplying a
-			 * component as a child of `Slider`, which follows the knob.
-			 *
-			 * @type {Boolean}
-			 * @default false
-			 * @public
-			 */
-			noTooltip: PropTypes.bool,
-
-			/**
 			 * The handler when the knob is activated or deactivated by selecting it via 5-way
 			 *
 			 * @type {Function}
@@ -238,6 +229,17 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			step: PropTypes.number,
 
 			/**
+			 * Enables the built-in tooltip, whose behavior can be modified by the other tooltip
+			 * properties.  A custom tooltip, which follows the knob, may be used instead by
+			 * supplying a component as a child of `Slider`.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			tooltip: PropTypes.bool,
+
+			/**
 			 * Converts the contents of the built-in tooltip to a percentage of the bar.
 			 * The percentage respects the min and max value props.
 			 *
@@ -297,10 +299,10 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			detachedKnob: false,
 			max: 100,
 			min: 0,
-			noTooltip: false,
 			onChange: () => {}, // needed to ensure the base input element is mutable if no change handler is provided
 			pressed: false,
 			step: 1,
+			tooltip: false,
 			tooltipAsPercent: false,
 			tooltipForceSide: false,
 			tooltipSide: 'before',
@@ -342,9 +344,9 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 		},
 
 		computed: {
-			children: ({children, max, min, noTooltip, tooltipAsPercent, value}) => {
+			children: ({children, max, min, tooltip, tooltipAsPercent, value}) => {
 				// If there's no tooltip, or custom children present, supply those.
-				if (noTooltip || children) return children;
+				if (!tooltip || children) return children;
 				return tooltipAsPercent ? Math.floor(computeProportionProgress({value, max, min}) * 100) + '%' : value;
 			},
 			className: ({active, pressed, vertical, styler}) => styler.append({
@@ -356,7 +358,7 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			proportionProgress: computeProportionProgress
 		},
 
-		render: ({backgroundProgress, children, disabled, inputRef, max, min, noTooltip, onBlur, onChange, onKeyDown, onMouseMove, onMouseUp, proportionProgress, scrubbing, sliderBarRef, sliderRef, step, tooltipForceSide, tooltipSide, value, vertical, ...rest}) => {
+		render: ({backgroundProgress, children, disabled, inputRef, max, min, onBlur, onChange, onKeyDown, onMouseMove, onMouseUp, proportionProgress, scrubbing, sliderBarRef, sliderRef, step, tooltip, tooltipForceSide, tooltipSide, value, vertical, ...rest}) => {
 			delete rest.active;
 			delete rest.detachedKnob;
 			delete rest.onActivate;
@@ -383,7 +385,7 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 						vertical={vertical}
 						scrubbing={scrubbing}
 					>
-						{noTooltip ? children : <SliderTooltip
+						{tooltip ? <SliderTooltip
 							className={css.tooltip}
 							forceSide={tooltipForceSide}
 							proportion={proportionProgress}
@@ -391,7 +393,7 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 							vertical={vertical}
 						>
 							{children}
-						</SliderTooltip>}
+						</SliderTooltip> : children}
 					</SliderBar>
 					<input
 						aria-disabled={disabled}
@@ -417,20 +419,28 @@ const SliderFactory = factory(css => {
 	const Base = SliderBaseFactory(css);
 
 	/**
-	 * {@link moonstone/Slider.Slider} is a Slider with Moonstone styling, Spottable, Pressable and
-	 * SliderDecorator applied. It is a stateful Slider.
+	 * {@link moonstone/Slider.Slider} is a Slider with Moonstone styling, Spottable, Changeable,
+	 * Pressable and SliderDecorator applied.
+	 *
+	 * By default, `Slider` maintains the state of its `value` property. Supply the `defaultValue`
+	 * property to control its initial value. If you wish to directly control updates to the
+	 * component, supply a value to `value` at creation time and update it in response to `onChange`
+	 * events.
 	 *
 	 * @class Slider
 	 * @memberof moonstone/Slider
-	 * @mixes spotlight/Spottable
-	 * @mixes ui/Pressable
+	 * @mixes ui/Changeable.Changeable
+	 * @mixes ui/Pressable.Pressable
+	 * @mixes spotlight/Spottable.Spottable
 	 * @ui
 	 * @public
 	 */
-	return Pressable(
-		Spottable(
-			SliderDecorator(
-				Base
+	return Changeable(
+		Pressable(
+			Spottable(
+				SliderDecorator(
+					Base
+				)
 			)
 		)
 	);