+++ packages/moonstone/Switch/Switch.js
@@ -52,7 +52,7 @@ const SwitchBase = kind({
 		 * @param {*} event.value - Value passed from `value` prop.
 		 * @public
 		 */
-		onToggle: PropTypes.bool,
+		onToggle: PropTypes.func,
 
 		/**
 		 * Sets whether this control is in the "on" or "off" state. `true` for on, `false` for "off".