+++ packages/sampler/stories/qa-stories/Button.js
@@ -1,12 +1,8 @@
-import Button, {ButtonBase} from '@enact/moonstone/Button';
+import Button from '@enact/moonstone/Button';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
 import {withKnobs, boolean, select, text} from '@kadira/storybook-addon-knobs';
 import css from './Button.less';
-
-Button.propTypes = Object.assign({}, ButtonBase.propTypes, Button.propTypes);
-Button.defaultProps = Object.assign({}, ButtonBase.defaultProps, Button.defaultProps);
-Button.displayName = 'Button';
 
 // Set up some defaults for info and knobs
 const prop = {