+++ packages/spotlight/src/spotlight.js
@@ -651,11 +651,15 @@ const Spotlight = (function () {
 	}
 
 	function getContainerLastFocusedElement (containerId) {
-		const lastFocusedElement = _containers.get(containerId).lastFocusedElement;
-		if (!isNavigable(lastFocusedElement, containerId, true)) {
-			return null;
+		const {lastFocusedElement, lastFocusedIndex} = _containers.get(containerId);
+
+		let element = lastFocusedElement;
+		if (!element && lastFocusedIndex >= 0) {
+			const spottableChildren = getContainerNavigableElements(containerId);
+			element = spottableChildren[lastFocusedIndex];
 		}
-		return lastFocusedElement;
+
+		return isNavigable(element, containerId, true) ? element : null;
 	}
 
 	function setContainerLastFocusedElement (elem, containerIds) {
@@ -1159,14 +1163,15 @@ const Spotlight = (function () {
 		 * @public
 		 */
 		set: function () {
-			let containerId, config;
+			let containerId, config, existingConfig;
 
 			if (typeof arguments[0] === 'object') {
 				config = arguments[0];
 			} else if (typeof arguments[0] === 'string' && typeof arguments[1] === 'object') {
 				containerId = arguments[0];
 				config = arguments[1];
-				if (!_containers.get(containerId)) {
+				existingConfig = _containers.get(containerId);
+				if (!existingConfig) {
 					throw new Error('Container "' + containerId + '" doesn\'t exist!');
 				}
 			} else {
@@ -1174,18 +1179,18 @@ const Spotlight = (function () {
 			}
 
 			for (let key in config) {
-				const validKey = typeof GlobalConfig[key] !== 'undefined';
-
-				if (!containerId && validKey && typeof config[key] !== 'undefined') {
-					GlobalConfig[key] = config[key];
-				} else if (containerId && !validKey) {
-					delete config[key];
+				if (typeof GlobalConfig[key] !== 'undefined') {
+					if (containerId) {
+						existingConfig[key] = config[key];
+					} else if (typeof config[key] !== 'undefined') {
+						GlobalConfig[key] = config[key];
+					}
 				}
 			}
 
 			if (containerId) {
 				// remove "undefined" items
-				_containers.set(containerId, extend({}, config));
+				_containers.set(containerId, extend({}, existingConfig));
 			}
 		},
 
@@ -1207,24 +1212,44 @@ const Spotlight = (function () {
 
 			if (typeof arguments[0] === 'object') {
 				config = arguments[0];
-			} else if (typeof arguments[0] === 'string' && typeof arguments[1] === 'object') {
+			} else if (typeof arguments[0] === 'string') {
 				containerId = arguments[0];
-				config = arguments[1];
+				if (typeof arguments[1] === 'object') {
+					config = arguments[1];
+				}
 			}
 
 			if (!containerId) {
 				containerId = (typeof config.id === 'string') ? config.id : generateId();
 			}
 
-			if (_containers.get(containerId)) {
-				throw new Error('Container "' + containerId + '" has already existed!');
+			// if a previous config does not exist, initialize a new one
+			if (!_containers.get(containerId)) {
+				_containers.set(containerId, config);
 			}
 
-			_containers.set(containerId, config);
-
+			// Either initialize the default config or merge the passed config with the existing
 			Spotlight.set(containerId, config);
 
 			return containerId;
+		},
+
+		unmount: function (containerId) {
+			if (!containerId || typeof containerId !== 'string') {
+				throw new Error('Please assign the "containerId"!');
+			}
+			const cfg = _containers.get(containerId);
+			if (cfg) {
+				const {lastFocusedElement} = cfg;
+				if (lastFocusedElement) {
+					const spottableChildren = getContainerNavigableElements(containerId);
+					const lastFocusedIndex = [].slice.call(spottableChildren).indexOf(lastFocusedElement);
+
+					// store last index and release node reference to lastFocusedElement
+					cfg.lastFocusedIndex = lastFocusedIndex;
+					cfg.lastFocusedElement = null;
+				}
+			}
 		},
 
 		/**